requirements.txt
python-jobspy>=1.1.75
requests>=2.31.0
beautifulsoup4>=4.12.3
algoliasearch>=3.0.0,<4.0.0
feedparser>=6.0.11
pandas>=2.1.0
pyyaml>=6.0.1
google-generativeai>=0.5.0
lxml>=5.1.0
pytest>=7.4.0
